+++ src/cookie/builder.rs
@@ -25,8 +25,8 @@ impl CookieBuilder {
     ///     .http_only()
     ///     .build();
     ///
-    /// assert!(cookie.secure());
-    /// assert!(cookie.http_only());
+    /// assert!(cookie.is_secure());
+    /// assert!(cookie.is_http_only());
     /// ```
     #[inline]
     pub fn new<N, V>(name: N, value: V) -> CookieBuilder
@@ -274,7 +274,7 @@ impl CookieBuilder {
     ///     .secure()
     ///     .build();
     ///
-    /// assert!(cookie.secure());
+    /// assert!(cookie.is_secure());
     /// ```
     #[inline]
     pub fn secure(mut self) -> Self {
@@ -299,7 +299,7 @@ impl CookieBuilder {
     ///     .http_only()
     ///     .build();
     ///
-    /// assert!(cookie.http_only());
+    /// assert!(cookie.is_http_only());
     /// ```
     #[inline]
     pub fn http_only(mut self) -> Self {
@@ -327,7 +327,7 @@ impl CookieBuilder {
     ///     .partitioned()
     ///     .build();
     ///
-    /// assert!(cookie.partitioned());
+    /// assert!(cookie.is_partitioned());
     /// ```
     #[inline]
     pub fn partitioned(self) -> Self {