+++ src/cookie/mod.rs
@@ -9,6 +9,7 @@ mod domain;
 pub(crate) mod expires;
 mod parse;
 mod path;
+mod prefix;
 pub(crate) mod same_site;
 mod serialize;
 
@@ -118,8 +119,8 @@ impl Cookie {
     ///     .http_only()
     ///     .build();
     ///
-    /// assert!(cookie.secure());
-    /// assert!(cookie.http_only());
+    /// assert!(cookie.is_secure());
+    /// assert!(cookie.is_http_only());
     /// ```
     pub fn build<N, V>(name: N, value: V) -> CookieBuilder
     where
@@ -257,7 +258,7 @@ impl Cookie {
 
     /// Returns if the Secure attribute is set.
     #[inline]
-    pub fn secure(&self) -> bool {
+    pub fn is_secure(&self) -> bool {
         self.secure
     }
 
@@ -269,7 +270,7 @@ impl Cookie {
 
     /// Returns if the HttpOnly attribute is set.
     #[inline]
-    pub fn http_only(&self) -> bool {
+    pub fn is_http_only(&self) -> bool {
         self.http_only
     }
 
@@ -281,7 +282,7 @@ impl Cookie {
 
     /// Returns if the Partitioned attribute is set.
     #[inline]
-    pub fn partitioned(&self) -> bool {
+    pub fn is_partitioned(&self) -> bool {
         self.partitioned
     }
 
@@ -320,9 +321,9 @@ impl Debug for Cookie {
             .field("max_age", &self.max_age())
             .field("domain", &self.domain())
             .field("path", &self.path())
-            .field("secure", &self.secure())
-            .field("http_only", &self.http_only())
-            .field("partitioned", &self.partitioned())
+            .field("secure", &self.is_secure())
+            .field("http_only", &self.is_http_only())
+            .field("partitioned", &self.is_partitioned())
             .field("expires", &self.expires)
             .field("same_site", &self.same_site)
             .finish()
@@ -333,9 +334,9 @@ impl PartialEq<Cookie> for Cookie {
     fn eq(&self, other: &Cookie) -> bool {
         if self.name() != other.name()
             || self.value() != other.value()
-            || self.secure() != other.secure()
-            || self.http_only() != other.http_only()
-            || self.partitioned() != other.partitioned()
+            || self.is_secure() != other.is_secure()
+            || self.is_http_only() != other.is_http_only()
+            || self.is_partitioned() != other.is_partitioned()
             || self.max_age() != other.max_age()
             || self.same_site() != other.same_site()
             || self.expires != other.expires